refactor(#55): extrait la construction de couches vers core/layer_builder.py
catalog.py reste limité au parsing du CSV (revue de PR, issue #52).
Comportement inchangé : contrats identiques, tests déplacés vers
test_layers.py, 135 tests verts sur QGIS 3.44.
