+++ specs/001-sharepoint-share-urls/plan.md
@@ -90,7 +90,10 @@ core/
 ├── __init__.py          # nouveau package de logique pure (aucun import qgis)
 ├── ms_urls.py           # is_microsoft_url(), is_sharepoint_sharing_link(),
 │                        # sharing_link_to_graph_url()
-└── catalog.py           # parsing/validation du CSV de flux (parse_table_row, catégories)
+├── catalog.py           # parsing/validation du CSV de flux (parse_table_row, catégories)
+└── layer_builder.py     # construction des URI de couches et périmètre sécurisé CEN
+                         # (build_wms_uri, build_wfs_uri_params, is_cen_secured_service) —
+                         # extrait de catalog.py en revue de PR (issue #52)
 
 config/yaml/
 └── links_example.yaml   # documentation mise à jour : lien de partage accepté tel quel
